Now handles local files, not just pecl

+++ scripts/pecl2ini/pecl2ini.php
@@ -24,30 +24,37 @@
 $pecl = $argv[1];
 
 print "Running for '$pecl'\n";
-$http = file_get_contents('http://pecl.php.net/package/'.$pecl);
-if (empty($http)) {
-    print "Can't load http://pecl.php.net/package/$pecl. Aborting\n";
-    die();
-}
-preg_match_all('#href="(/get/('.$pecl.'-[0-9RC\.]+\.tgz))"#is', $http, $versions);
-
-//print_r($versions);
-$version = $versions[2][0];
-
-if (!file_exists($version)) {
-    file_put_contents($version, file_get_contents("http://pecl.php.net".$versions[1][0]));
-    if (!file_exists($version)) {
-        print "Can't load source code for {$versions[1][0]}. Aborting\n";
+if (file_exists("./$pecl")) {
+    $path = "./$pecl";
+} else {
+    $http = file_get_contents("http://pecl.php.net/package/$pecl");
+    if (empty($http)) {
+        print "Can't load http://pecl.php.net/package/$pecl. Aborting\n";
         die();
     }
-} else {
-    print "$version est deja ici\n";
+    preg_match_all('#href="(/get/('.$pecl.'-[0-9RC\.]+\.tgz))"#is', $http, $versions);
+
+    //print_r($versions);
+    $version = $versions[2][0];
+
+    if (!file_exists($version)) {
+        file_put_contents($version, file_get_contents("http://pecl.php.net".$versions[1][0]));
+        if (!file_exists($version)) {
+            print "Can't load source code for {$versions[1][0]}. Aborting\n";
+            die();
+        }
+    } else {
+        print "$version est deja ici\n";
+    }
+    shell_exec('tar -xf '.$version);
+    
+    $path = $version;
 }
 
-shell_exec('tar -xf '.$version);
+
 
 // functions
-$res = shell_exec('grep -r PHP_FUNCTION '.substr($version, 0, -4));
+$res = shell_exec('grep -r PHP_FUNCTION '.$path);
 if (empty($res)) {
     $classes = array(); 
     print "No functions\n\n";
@@ -60,9 +67,9 @@ if (empty($res)) {
 }
 
 // functions
-$res = shell_exec('grep -r INIT_CLASS_ENTRY '.substr($version, 0, -4));
+$res = shell_exec('grep -r INIT_CLASS_ENTRY '.$path);
 if (empty($res)) {
-    $res = shell_exec('grep -r PHP_EVENT_REGISTER_CLASS '.substr($version, 0, -4));
+    $res = shell_exec('grep -r PHP_EVENT_REGISTER_CLASS '.$path);
     if (empty($res)) {
         $classes = array(); 
         print "No classes\n\n";
@@ -72,29 +79,32 @@ if (empty($res)) {
         print count($classes)." classes\n  ";
         print implode("\n  ", $classes);
         print "\n\n";
-    
     }
 } elseif (preg_match_all('/INIT_CLASS_ENTRY\([a-z_]+, "([a-zA-Z0-9_]+)"/is', $res, $classes)) {
     $classes = $classes[1];
     print count($classes)." classes\n  "; 
     print implode("\n  ", $classes);
     print "\n\n";
-} elseif (preg_match_all('/INIT_CLASS_ENTRY\([a-z_]+,\s*([a-zA-Z0-9_]+)/is', $res, $classesDefine)) {
+} elseif (preg_match_all('/INIT_CLASS_ENTRY\([a-z\\_]+,\s*([a-zA-Z0-9_]+)/is', $res, $classesDefine)) {
     $classesDefine = $classesDefine[1];
-    $res = shell_exec('grep -r "('.implode('\\|', $classesDefine).')" '.substr($version, 0, -4));
+    $res = shell_exec('grep -r "('.implode('\\|', $classesDefine).')" '.$path);
     preg_match_all("/(".implode('|', $classesDefine).") \"([a-zA-Z0-9_]+)\"/is", $res, $classes);
     $classes = $classes[2];
     print count($classes)." classes\n  ";
     print implode("\n  ", $classes);
+    print "\n\n";
+} elseif (preg_match_all('/INIT_CLASS_ENTRY\(ce,\s*"([a-z\\\\_]+)",\s*([a-zA-Z0-9_]+)/is', $res, $classesDefine)) {
+    print count($classesDefine[1])." classes\n  ";
+    print implode("\n  ", $classesDefine[1]);
     print "\n\n";
 } else {
     print "Nothing found, but having res : $res\n";
 }
 
 // constants
-$res = shell_exec('grep -r REGISTER_CAIRO_STATUS_LONG_CONST '.substr($version, 0, -4));
+$res = shell_exec('grep -r REGISTER_CAIRO_STATUS_LONG_CONST '.$path);
 if (empty($res)) {
-$res = shell_exec('grep -r REGISTER_LONG_CONSTANT '.substr($version, 0, -4));
+$res = shell_exec('grep -r REGISTER_LONG_CONSTANT '.$path);
     if (empty($res)) {  
         $constants = array(); 
         print "No constants\n\n";
